shopperstack-backend/requirements.txt
fastapi
uvicorn[standard]
uagents
requests
python-dotenv
tavily-python
